requirements.txt
langchain
langchain-community
chromadb
fastapi
uvicorn
pydantic
python-dotenv
langchain-openai
chardet